Widen animation file offsets to 32 bits.
Frame offsets are 28 + 315*frame, which exceeds a uint16_t at frame 208;
the old signature silently truncated the seek and read the wrong frame.
spiffs seeks are signed 32-bit, so uint32_t covers every reachable offset.

+++ ccs_workspace/badge.lgbt/badge_drivers/storage.h
@@ -19,7 +19,7 @@ extern uint16_t storage_next_anim_id;
 extern char storage_anim_id_cache[STORAGE_ANIMS_TO_CACHE][ANIM_NAME_MAX_LEN];
 
 void storage_init();
-uint8_t storage_read_file(char *fname, uint8_t *dest, uint16_t offset, uint16_t size);
+uint8_t storage_read_file(char *fname, uint8_t *dest, uint32_t offset, uint16_t size);
 uint8_t storage_load_frame(char *anim_name, uint16_t frame_number, rgbcolor_t (*dest)[15]);
 void storage_save_direct_anim(char *anim_name, led_anim_direct_t *anim, uint8_t unlocked);
 uint8_t storage_anim_saved_and_valid(char *anim_name);
